refactor: 싱글턴 테스트 클래스 스태틱으로 변경
